+++ x/blob/types/payforblob.go
@@ -1,20 +1,18 @@
 package types
 
 import (
-	"crypto/sha256"
 	fmt "fmt"
 
 	"cosmossdk.io/errors"
 
 	"github.com/celestiaorg/celestia-app/pkg/appconsts"
 	"github.com/celestiaorg/celestia-app/pkg/blob"
+	"github.com/celestiaorg/celestia-app/pkg/inclusion"
 	appns "github.com/celestiaorg/celestia-app/pkg/namespace"
 	appshares "github.com/celestiaorg/celestia-app/pkg/shares"
-	"github.com/celestiaorg/nmt"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 	auth "github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/tendermint/tendermint/crypto/merkle"
 	"golang.org/x/exp/slices"
 )
 
@@ -51,7 +49,7 @@ func NewMsgPayForBlobs(signer string, blobs ...*blob.Blob) (*MsgPayForBlobs, err
 	if err != nil {
 		return nil, err
 	}
-	commitments, err := CreateCommitments(blobs)
+	commitments, err := inclusion.CreateCommitments(blobs)
 	if err != nil {
 		return nil, err
 	}
@@ -209,81 +207,6 @@ func (msg *MsgPayForBlobs) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{address}
 }
 
-// CreateCommitment generates the share commitment for a given blob.
-// See [data square layout rationale] and [blob share commitment rules].
-//
-// [data square layout rationale]: ../../specs/src/specs/data_square_layout.md
-// [blob share commitment rules]: ../../specs/src/specs/data_square_layout.md#blob-share-commitment-rules
-func CreateCommitment(blob *blob.Blob) ([]byte, error) {
-	if err := blob.Validate(); err != nil {
-		return nil, err
-	}
-	namespace := blob.Namespace()
-
-	shares, err := appshares.SplitBlobs(blob)
-	if err != nil {
-		return nil, err
-	}
-
-	// the commitment is the root of a merkle mountain range with max tree size
-	// determined by the number of roots required to create a share commitment
-	// over that blob. The size of the tree is only increased if the number of
-	// subtree roots surpasses a constant threshold.
-	subTreeWidth := appshares.SubTreeWidth(len(shares), appconsts.DefaultSubtreeRootThreshold)
-	treeSizes, err := MerkleMountainRangeSizes(uint64(len(shares)), uint64(subTreeWidth))
-	if err != nil {
-		return nil, err
-	}
-	leafSets := make([][][]byte, len(treeSizes))
-	cursor := uint64(0)
-	for i, treeSize := range treeSizes {
-		leafSets[i] = appshares.ToBytes(shares[cursor : cursor+treeSize])
-		cursor = cursor + treeSize
-	}
-
-	// create the commitments by pushing each leaf set onto an nmt
-	subTreeRoots := make([][]byte, len(leafSets))
-	for i, set := range leafSets {
-		// create the nmt todo(evan) use nmt wrapper
-		tree := nmt.New(sha256.New(), nmt.NamespaceIDSize(appns.NamespaceSize), nmt.IgnoreMaxNamespace(true))
-		for _, leaf := range set {
-			// the namespace must be added again here even though it is already
-			// included in the leaf to ensure that the hash will match that of
-			// the nmt wrapper (pkg/wrapper). Each namespace is added to keep
-			// the namespace in the share, and therefore the parity data, while
-			// also allowing for the manual addition of the parity namespace to
-			// the parity data.
-			nsLeaf := make([]byte, 0)
-			nsLeaf = append(nsLeaf, namespace.Bytes()...)
-			nsLeaf = append(nsLeaf, leaf...)
-
-			err = tree.Push(nsLeaf)
-			if err != nil {
-				return nil, err
-			}
-		}
-		// add the root
-		root, err := tree.Root()
-		if err != nil {
-			return nil, err
-		}
-		subTreeRoots[i] = root
-	}
-	return merkle.HashFromByteSlices(subTreeRoots), nil
-}
-
-func CreateCommitments(blobs []*blob.Blob) ([][]byte, error) {
-	commitments := make([][]byte, len(blobs))
-	for i, blob := range blobs {
-		commitment, err := CreateCommitment(blob)
-		if err != nil {
-			return nil, err
-		}
-		commitments[i] = commitment
-	}
-	return commitments, nil
-}
-
 // ValidateBlobs performs basic checks over the components of one or more PFBs.
 func ValidateBlobs(blobs ...*blob.Blob) error {
 	if len(blobs) == 0 {
@@ -331,31 +254,4 @@ func ExtractBlobComponents(pblobs []*blob.Blob) (namespaceVersions []uint32, nam
 	}
 
 	return namespaceVersions, namespaceIds, sizes, shareVersions
-}
-
-// MerkleMountainRangeSizes returns the sizes (number of leaf nodes) of the
-// trees in a merkle mountain range constructed for a given totalSize and
-// maxTreeSize.
-//
-// https://docs.grin.mw/wiki/chain-state/merkle-mountain-range/
-// https://github.com/opentimestamps/opentimestamps-server/blob/master/doc/merkle-mountain-range.md
-func MerkleMountainRangeSizes(totalSize, maxTreeSize uint64) ([]uint64, error) {
-	var treeSizes []uint64
-
-	for totalSize != 0 {
-		switch {
-		case totalSize >= maxTreeSize:
-			treeSizes = append(treeSizes, maxTreeSize)
-			totalSize = totalSize - maxTreeSize
-		case totalSize < maxTreeSize:
-			treeSize, err := appshares.RoundDownPowerOfTwo(totalSize)
-			if err != nil {
-				return treeSizes, err
-			}
-			treeSizes = append(treeSizes, treeSize)
-			totalSize = totalSize - treeSize
-		}
-	}
-
-	return treeSizes, nil
 }
